requirements.txt
setuptools
pandas
numpy

stable-audio-tools
torchaudio

alias-free-torch
auraloss
descript-audio-codec
descript-audiotools
einops-exts
ema-pytorch
k-diffusion
laion-clap
local-attention

argbind
clean-fid
clip-anytorch
dctorch
jsonmerge
transformers
accelerate